Added: All ejs files and logic to output users
Added: modified the server to use the EJS template engine and render the added EJS versions of the server files added to the project.

Added the logic to display the output from the form on the root route of the project on the users route showing the entered usernames. using Objects to pass the data to EJS.

NOTE: EJS and Pug(Jade) can both have data passed to them via objects since this is native to Node.js

Assignment  Objective completed.